# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/site": {
            "put": {
                "description": "站点相关接口",
                "tags": [
                    "站点管理"
                ],
                "summary": "更新站点信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "分组信息",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dao.UpdateSite"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"msg\": \"更新成功\", \"data\": nil}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "站点相关接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "站点管理"
                ],
                "summary": "创建站点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "分组信息",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.SiteCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"msg\": \"创建成功\", \"data\": nil}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/site/group": {
            "put": {
                "description": "站点相关接口",
                "tags": [
                    "站点管理"
                ],
                "summary": "更新分组信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "分组信息",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.SiteGroupUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"msg\": \"更新成功\", \"data\": nil}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "站点相关接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "站点管理"
                ],
                "summary": "创建分组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "分组信息",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.SiteGroupCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"msg\": \"创建成功\", \"data\": nil}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/site/group/{id}": {
            "delete": {
                "description": "站点相关接口",
                "tags": [
                    "站点管理"
                ],
                "summary": "删除站点分组",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"msg\": \"删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/site/guide": {
            "get": {
                "description": "站点相关接口",
                "tags": [
                    "站点管理"
                ],
                "summary": "获取站点列表（导航）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": []}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/site/logoUpload": {
            "post": {
                "description": "站点相关接口",
                "tags": [
                    "站点管理"
                ],
                "summary": "站点图片上传",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "头像",
                        "name": "logo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"path\": logoPath}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/site/tags": {
            "put": {
                "description": "站点相关接口",
                "tags": [
                    "站点管理"
                ],
                "summary": "更新站点标签",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "users",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.SiteTagUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"msg\": \"更新成功\", \"data\": nil}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/site/users": {
            "put": {
                "description": "站点相关接口",
                "tags": [
                    "站点管理"
                ],
                "summary": "更新站点用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "users",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.SiteUserUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"msg\": \"更新成功\", \"data\": nil}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/site/{id}": {
            "delete": {
                "description": "站点相关接口",
                "tags": [
                    "站点管理"
                ],
                "summary": "删除站点",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分组ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"msg\": \"删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/sites": {
            "get": {
                "description": "站点相关接口",
                "tags": [
                    "站点管理"
                ],
                "summary": "获取站点列表（表格）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分页",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分页大小",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "站点名称",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": []}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dao.UpdateSite": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "all_open": {
                    "description": "指针类型，可以确保使用Updates方法更新时，如果值为false时也能更新成功",
                    "type": "boolean"
                },
                "callback_url": {
                    "type": "string"
                },
                "certificate": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "entity_id": {
                    "description": "指针类型，可以确保使用Updates方法更新时，如果值为空时也能更新成功",
                    "type": "string"
                },
                "helper_url": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sso": {
                    "description": "指针类型，可以确保使用Updates方法更新时，如果值为false时也能更新成功",
                    "type": "boolean"
                },
                "sso_type": {
                    "type": "integer"
                }
            }
        },
...
        "service.SiteCreate": {
            "type": "object",
            "required": [
                "address",
                "description",
                "name",
                "site_group_id",
                "sso"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "callback_url": {
                    "type": "string"
                },
                "certificate": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "domain_id": {
                    "type": "string"
                },
                "entity_id": {
                    "type": "string"
                },
                "helper_url": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "idp_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "redirect_url": {
                    "type": "string"
                },
                "site_group_id": {
                    "type": "integer"
                },
                "sso": {
                    "type": "boolean"
                },
                "sso_type": {
                    "type": "integer"
                }
            }
        },
        "service.SiteGroupCreate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "service.SiteGroupUpdate": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "service.SiteTagUpdate": {
            "type": "object",
            "required": [
                "id",
                "tags"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "service.SiteUserUpdate": {
            "type": "object",
            "required": [
                "id",
                "users"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}